src/requirements.txt
Flask>=2.0
Flask-Cors>=3.0
Flask-SQLAlchemy>=2.5
SQLAlchemy>=1.4
marshmallow>=3.0
pymssql>=2.2
python-dotenv>=0.21
PyJWT>=2.0
alembic>=1.8
flet>=0.9.0
Werkzeug>=2.0